Refactor: move `XFormsProperties` to Scala and Scala-ify

+++ src/main/scala/org/orbeon/oxf/util/SecureUtils.scala
@@ -42,7 +42,7 @@ object SecureUtils {
   }
 
   private def getKeyLength: Int =
-    Properties.instance.getPropertySet.getInteger(KeyLengthProperty, 128).intValue
+    Properties.instance.getPropertySet.getInteger(KeyLengthProperty, 128)
 
   private def getHashAlgorithm: String =
     Properties.instance.getPropertySet.getString(HashAlgorithmProperty, "SHA1")
